requirements.txt
dbt-core
dbt-postgres
minio
apache-airflow